pyproject.toml
[project]
name = "contabilidad-iconstruye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "selenium>=4.38.0",
    "xlsx2csv>=0.8.4",
    "requests>=2.32.5",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pydrive2>=1.15.0",
    "pydantic>=2.12.3",
    "playwright>=1.55.0",
    "google-auth>=2.32.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.185.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]